server/server/views.py
from rest_framework import viewsets
from server.models import Todo
from server.serializers import TodoSerializer


class TodoViewSet(viewsets.ModelViewSet):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer
