server/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
django-filter==1.1.0
django-cors-headers==2.4.0